압축 파일 해제 기능 개선

+++ 4.Tool/CLI/handstack/Program.cs
@@ -842,12 +842,12 @@ namespace handstack
                 {
                     if (file != null && file.Exists == true && directory != null)
                     {
-                        if (directory.Exists == true)
+                        if (directory.Exists == true && key.ToBoolean() == true)
                         {
                             directory.Delete(true);
                         }
 
-                        ZipFile.ExtractToDirectory(file.FullName, directory.FullName);
+                        ZipFile.ExtractToDirectory(file.FullName, directory.FullName, true);
                     }
                     else
                     {
